requirements.txt
attrs==22.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.0.4
Flask==2.2.2
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
tomli==2.0.1
Werkzeug==2.2.2
